refactor: use long for profile stats

+++ backend/src/main/java/format/backend/auth/dto/UserStatisticsDto.java
@@ -3,6 +3,6 @@ package format.backend.auth.dto;
 import org.jspecify.annotations.NonNull;
 
 public record UserStatisticsDto(
-        @NonNull Integer formsCount,
-        @NonNull Integer submissionsCount,
-        @NonNull Integer commentsCount) {}
+        @NonNull Long formsCount,
+        @NonNull Long submissionsCount,
+        @NonNull Long commentsCount) {}
